+++ activerecord/lib/active_record/railtie.rb
@@ -454,5 +454,18 @@ To keep using the current cache store, you can turn off cache versioning entirel
         end
       end
     end
+
+    initializer "active_record.freeze_configuration" do |app|
+      config.after_initialize do
+        if app.config.freeze_configuration
+          ActiveSupport::Ractors.make_shareable(ActiveRecord.protected_environments)
+          ActiveSupport::Ractors.make_shareable(ActiveRecord.schema_cache_ignored_tables)
+          ActiveSupport::Ractors.make_shareable(ActiveRecord.database_cli)
+          ActiveSupport::Ractors.make_shareable(ActiveRecord.db_warnings_ignore)
+          ActiveSupport::Ractors.make_procs_shareable(ActiveRecord.query_transformers)
+          ActiveSupport::Ractors.make_shareable(ActiveRecord.yaml_column_permitted_classes)
+        end
+      end
+    end
   end
 end